Fix public metadata review findings
