+++ log.md
@@ -14,3 +14,4 @@
 - [2026-07-29] Cloudflare cache-tag rollout: `Cache-Tag` headers do not tag objects that Cloudflare already cached. When a release introduces tag-based deploy purges, purge the hostname once after the tagged release reaches origin; later deploys can purge only the stable tag.
 - [2026-07-30] Ahrefs representative crawl scope: A custom-list-only crawl with maximum depth 0 reports sitemap and orphan warnings because sitemap sources and link traversal are disabled. Use this crawl for route-template checks, and validate sitemap membership separately.
 - [2026-07-30] Prod-like crawl analytics isolation: `prepare-prod-like-local-smoke.mjs` disables Sentry but retains the production PostHog configuration. Before a crawler uses `dev.daylilycatalog.com`, set `NEXT_PUBLIC_POSTHOG_KEY: ""` in the ignored prod-like Compose override and verify `/api/runtime-config` reports PostHog disabled.
+- [2026-07-31] Pure metadata test imports: Importing a database-backed App Router module can trigger environment validation before Vitest executes a test, while local environment files can hide the problem. Keep pure metadata builders in database-free modules and import those modules directly in focused tests.
